Graduate MaskTransform from prototype
