+++ torchtext/transforms.py
@@ -1,7 +1,7 @@
 import json
 from copy import deepcopy
 from functools import lru_cache
-from typing import Any, List, Optional, Union
+from typing import Any, List, Optional, Tuple, Union
 
 import torch
 import torchtext  # noqa: F401
@@ -767,3 +767,144 @@ class Sequential(torch.nn.Sequential):
         for module in self:
             input = module(input)
         return input
+
+
+class MaskTransform(torch.nn.Module):
+    """
+    The transform chooses mask_prob% (example 15%) of the token positions at random for
+    prediction.
+
+    If the i-th token is chosen, we replace the i-th token with
+    (1) the [MASK] token 80% of the time
+    (2) a random token 10% of the time
+    (3) the unchanged i-th token 10% of the time.
+
+    Args:
+        vocab_len (int): the length of the vocabulary, including special tokens such as [BOS], [PAD], [MASK]
+        mask_idx (int): index assigned to mask token in vocabulary
+        bos_idx (int): index assigned to beginning-of-sequence token in vocabulary
+        pad_idx (int): index assigned to padding token in vocabulary
+        mask_bos (bool): indicate whether beginning-of-sequence tokens are eligible for masking (default: False)
+        mask_prob (float): probability that a token is chosen for replacement (default: 0.15)
+
+    Example:
+        >>> import torch
+        >>> from torchtext.transforms import MaskTransform
+        >>> sample_tokens = [
+                ["[BOS]", "a", "b", "c", "d"],
+                ["[BOS]", "a", "b", "[PAD]", "[PAD]"]
+            ]
+        >>> sample_token_ids = torch.tensor([
+                [6, 0, 1, 2, 3], [6, 0, 1, 4, 4]
+            ])
+        >>> mask_transform = MaskTransform(
+                vocab_len = 7,
+                mask_idx = 4,
+                bos_idx = 6,
+                pad_idx = 5,
+                mask_bos = False,
+                mask_prob = 0.15
+            )
+        >>> masked_tokens, target_tokens, mask = mask_transform(sample_token_ids)
+    """
+
+    # maks_mask_prob is prob. of replacing a token with [MASK] (ex. 80%)
+    mask_mask_prob = 0.8
+
+    # rand_mask_thresh is prob. of replacing a token with a random token. (ex.10%)
+    rand_mask_prob = 0.1
+
+    def __init__(
+        self,
+        vocab_len: int,
+        mask_idx: int,
+        bos_idx: int,
+        pad_idx: int,
+        mask_bos: bool = False,
+        mask_prob: float = 0.15,
+    ):
+        super().__init__()
+        self.vocab_len = vocab_len
+        self.mask_idx = mask_idx
+        self.bos_idx = bos_idx
+        self.pad_idx = pad_idx
+        self.mask_prob = mask_prob
+        self.mask_bos = mask_bos
+
+    def forward(self, tokens: torch.Tensor) -> Tuple[torch.Tensor, torch.Tensor, torch.Tensor]:
+        """
+        Applies mask to input tokens.
+
+        Inputs:
+            tokens: Tensor with token ids of shape (batch_size x seq_len). Includes token ids for special tokens such as [BOS] and [PAD]
+
+        Outputs:
+            masked_tokens: Tensor of tokens after masking has been applied
+            target_tokens: Tensor of token values selected for masking
+            mask: Tensor with same shape as input tokens (batch_size x seq_len)
+                with masked tokens represented by a 1 and everything else as 0.
+        """
+        # tokens, mask, mask_mask, rand_mask: (T, C)
+        mask, mask_mask, rand_mask = self._generate_mask(tokens)
+
+        # a. generate the masked input tokens
+        # (1) the [MASK] token 80% of the time
+        masked_tokens = self._mask_input(tokens, mask_mask, self.mask_idx)
+        # (2) a random token 10% of the time
+        masked_tokens = self._mask_input(
+            masked_tokens,
+            rand_mask,
+            torch.randint_like(tokens, high=self.vocab_len),
+        )
+
+        # b. generate the target prediction
+        target_tokens = torch.masked_select(tokens, mask.bool())
+
+        # masked_tokens: (T, C), target_tokens: (T x C x mask_prob, ), mask
+        return masked_tokens, target_tokens, mask
+
+    def _random_masking(self, tokens: torch.tensor, mask_prob: float) -> torch.Tensor:
+        """
+        Function to mask tokens randomly.
+
+        Inputs:
+            1) tokens: Tensor with token ids of shape (batch_size x seq_len). Includes token ids for special tokens such as [BOS] and [PAD]
+            2) mask_prob: Probability of masking a particular token
+
+        Outputs:
+            mask: Tensor with same shape as input tokens (batch_size x seq_len)
+                with masked tokens represented by a 1 and everything else as 0.
+        """
+        batch_size, seq_len = tokens.size()
+        num_masked_per_seq = int(seq_len * mask_prob)
+
+        mask = torch.zeros((batch_size, seq_len), dtype=torch.int).to(tokens.device)
+        mask[:, :num_masked_per_seq] = 1
+        for i in range(batch_size):
+            mask[i] = mask[i, torch.randperm(seq_len)]
+
+        return mask
+
+    def _select_tokens_to_mask(self, tokens: torch.Tensor, mask_prob: float) -> torch.Tensor:
+        mask = self._random_masking(tokens, mask_prob)
+        if not self.mask_bos:
+            mask *= (tokens != self.bos_idx).long()
+        return mask
+
+    def _generate_mask(self, tokens: torch.Tensor) -> Tuple[torch.Tensor, torch.Tensor, torch.Tensor]:
+        # chooses mask_prob% of the token positions at random
+        mask = self._select_tokens_to_mask(tokens, self.mask_prob)
+        # not mask the pad token
+        mask *= (tokens != self.pad_idx).long()
+        # keep one masked token to avoid failure in the loss calculation.
+        mask[0, 0] = 1 if not mask.byte().any() else mask[0, 0]
+
+        probs = torch.rand_like(tokens, dtype=torch.float)
+        # (1) the [MASK] token 80% of the time
+        mask_mask = (probs >= (1 - self.mask_mask_prob)).long() * mask
+        # (2) a random token 10% of the time
+        rand_mask = (probs < self.rand_mask_prob).long() * mask
+        return mask, mask_mask, rand_mask
+
+    def _mask_input(self, tokens: torch.Tensor, mask: torch.Tensor, replacement) -> torch.Tensor:
+        return tokens * (1 - mask) + replacement * mask
